fix: add date validation

+++ main-service/src/main/java/ru/practicum/controller/AdminController.java
@@ -68,9 +68,9 @@ public class AdminController {
 
     @GetMapping("/events")
     @ResponseStatus(HttpStatus.OK)
-    public List<FullEventDto> getEvents(@RequestParam(required = false) List<Integer> users,
+    public List<FullEventDto> getEvents(@RequestParam(required = false) List<Long> users,
                                         @RequestParam(required = false) List<String> states,
-                                        @RequestParam(required = false) List<Integer> categories,
+                                        @RequestParam(required = false) List<Long> categories,
                                         @RequestParam(required = false) String rangeStart,
                                         @RequestParam(required = false) String rangeEnd,
                                         @RequestParam(defaultValue = "0") int from,
